Upgrades to .NET 10 and related packages (#2)
* Upgrades to .NET 10 and related packages

Updates the target framework to .NET 10 across the solution.

Also upgrades several dependency packages, including Microsoft.Extensions.*, NLog, NUnit, and Telegram.Bot to their latest versions.

These updates ensure compatibility with the newest .NET runtime and provide access to the latest features, bug fixes, and performance improvements in the upgraded packages.

* Updates .NET version to 10.0

Updates the .NET version used in the test workflow to 10.0.

This ensures that the tests are run against the latest .NET version.